Move get_site_for_user to wagtail.admin.navigation
