+++ wagtail/admin/utils.py
@@ -1,25 +1,6 @@
 import sys
 
-from django.conf import settings
-
-from wagtail.admin.navigation import get_explorable_root_page
 from wagtail.utils.deprecation import MovedDefinitionHandler, RemovedInWagtail29Warning
-
-
-def get_site_for_user(user):
-    root_page = get_explorable_root_page(user)
-    if root_page:
-        root_site = root_page.get_site()
-    else:
-        root_site = None
-    real_site_name = None
-    if root_site:
-        real_site_name = root_site.site_name if root_site.site_name else root_site.hostname
-    return {
-        'root_page': root_page,
-        'root_site': root_site,
-        'site_name': real_site_name if real_site_name else settings.WAGTAIL_SITE_NAME,
-    }
 
 
 MOVED_DEFINITIONS = {
@@ -41,6 +22,8 @@ MOVED_DEFINITIONS = {
 
     'send_mail': 'wagtail.admin.mail',
     'send_notification': 'wagtail.admin.mail',
+
+    'get_site_for_user': 'wagtail.admin.navigation',
 }
 
 sys.modules[__name__] = MovedDefinitionHandler(sys.modules[__name__], MOVED_DEFINITIONS, RemovedInWagtail29Warning)
